Move client_id_and_secret fixture
